Create ex 1
Task: A numeric array A[m] is given. Calculate and print the number of different numbers in this array. For example, in the array 5,7,5 there are 2 different numbers 7 and 5. Transform the array in such a way that all positive elements are placed first, and then all negative ones (elements equal to 0 are considered positive).